mainAPI/requirements.txt
pymongo~=4.11
Flask~=3.1.0
anthropic~=0.45.2
python-dotenv~=1.0.1
elevenlabs~=1.50.6